Divided the code into modules
